Crate clippy_lints

Modules

approx_const

arithmetic

array_indexing

assign_ops

attrs
checks for attributes

bit_mask

blacklisted_name

block_in_if_condition

booleans

collapsible_if
Checks for if expressions that contain only an if expression.

consts

copies

cyclomatic_complexity
calculate cyclomatic complexity and warn about overly complex functions

derive

doc

drop_ref

entry

enum_clike
lint on C-like enums that are repr(isize/usize) and have values that don't fit into an i32

enum_glob_use
lint on useing all variants of an enum

enum_variants
lint on enum variants that are prefixed or suffixed by the same characters

eq_op

escape

eta_reduction

eval_order_dependence

format

formatting

functions

identity_op

if_not_else
lint on if branches that could be swapped so no ! operation is necessary on the condition

items_after_statements
lint when items are used after statements

len_zero

let_if_seq

lifetimes

loops

map_clone

matches

mem_forget

methods

minmax

misc

misc_early

missing_doc

module_inception

mut_mut

mut_reference

mutex_atomic
Checks for uses of mutex where an atomic value could be used

needless_bool
Checks for needless boolean results of if-else expressions

needless_borrow
Checks for needless address of operations (&)

needless_update

neg_multiply

new_without_default

no_effect

non_expressive_names

open_options

overflow_check_conditional

panic

precedence

print

ptr
Checks for usage of &Vec[_] and &String.

ranges

regex

returns

serde

shadow

strings

swap

temporary_assignment

transmute

types

unicode

unsafe_removed_from_name

unused_label

utils

vec

zero_div_zero

Macros

if_let_chain
Produce a nested chain of if-lets and ifs from the patterns:

Functions

register_plugins