Function clippy_lints::doc::strip_doc_comment_decoration

pub fn strip_doc_comment_decoration((comment, span): (&str, Span)) -> Vec<(&str, Span)>

Cleanup documentation decoration (/// and such).

We can't use syntax::attr::AttributeMethods::with_desugared_doc or syntax::parse::lexer::comments::strip_doc_comment_decoration because we need to keep track of the span but this function is inspired from the later.