Struct clippy_lints::utils::conf::Conf

pub struct Conf {
    pub blacklisted_names: Vec<String>,
    pub cyclomatic_complexity_threshold: u64,
    pub doc_valid_idents: Vec<String>,
    pub too_many_arguments_threshold: u64,
    pub type_complexity_threshold: u64,
    pub max_single_char_names: u64,
    pub too_large_for_stack: u64,
    pub enum_variant_name_threshold: u64,
}

Type used to store lint configuration.

Fields
blacklisted_names: Vec<String>

Lint: BLACKLISTED_NAME. The list of blacklisted names to lint about

cyclomatic_complexity_threshold: u64

Lint: CYCLOMATIC_COMPLEXITY. The maximum cyclomatic complexity a function can have

doc_valid_idents: Vec<String>

Lint: DOC_MARKDOWN. The list of words this lint should not consider as identifiers needing ticks

too_many_arguments_threshold: u64

Lint: TOO_MANY_ARGUMENTS. The maximum number of argument a function or method can have

type_complexity_threshold: u64

Lint: TYPE_COMPLEXITY. The maximum complexity a type can have

max_single_char_names: u64

Lint: MANY_SINGLE_CHAR_NAMES. The maximum number of single char bindings a scope may have

too_large_for_stack: u64

Lint: BOXED_LOCAL. The maximum size of objects (in bytes) that will be linted. Larger objects are ok on the heap

enum_variant_name_threshold: u64

Lint: ENUM_VARIANT_NAMES. The minimum number of enum variants for the lints about variant names to trigger