Module clippy_lints::utils::higher

This module contains functions for retrieve the original AST from lowered hir.

Structs

Range
Represent a range akin to ast::ExprKind::Range.

Enums

VecArgs
Represent the pre-expansion arguments of a vec! invocation.

Functions

binop
Convert a hir binary operator to the corresponding ast type.

for_loop
Recover the essential nodes of a desugared for loop: for pat in arg { body } becomes (pat, arg, body).

is_from_for_desugar
Checks if a let decl is from a for loop desugaring.

range
Higher a hir range to something similar to ast::ExprKind::Range.

vec_macro
Returns the arguments of the vec! macro if this expression was expanded from vec!.